+++ changelog.rst
@@ -13,6 +13,7 @@ Moldy Bread Change Log
 
 * Get a list of objects created or modified by a user with audit_responsibility.  See docs for more details.
 * Find all objects matching your requested number of versions of a particular datastream.  See docs for more details.
+* Find all objects that have no management restrictions (deny-management-functions).  See docs for more details.
 
 **New Fedora Public Methods and Other Procs**: